apps/complete-physio-weekly-summary/requirements.txt
instaloader==4.10.3
feedparser==6.0.10
python-dateutil==2.8.2
schedule==1.2.0
pytz==2023.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
